fix: not all notes indexed in sparse vaults
- fixes https://github.com/dendronhq/dendron/issues/71

### Context
The way Dendron indexes notes is by dependent on the maximum depth (depth is dependent by number of `.` + 1) of any note

- eg:
    - foo -> depth 1
    - foo.bar.baz -> depth 3

During traversal, Dendron will fill any missing notes with stubs but it only traverses up to the maximum depth of a note hierarchy that has a full path.

- Dendron will travel two levels because foo.bar is the longest valid path
```
- foo.md
- foo.bar.md
```

- dendron will only travel one level because foo is the longest valid path
```
- foo.md
- foo.baz.bar.md
```

When initializing a new vault, your longest valid path is 0. If you create a note with stubs in between, dendron will not be able to index it.