Fix rejected promises in testing (#746)
* Fix incorrectly passing tests

* Fix failing DiagnosticReport test

The test was silently failing as the rejected promises were not being
checked. The actual bugs were that (possibly due to a change in VSCode
internals) the opened document was not getting focused, and the
incorrect reverse slicing was causing the last character of logs to be
lost.

* Fix silently failing MoveNoteCommand test

The test was silently failing because the promise returned by `lc.show`
was not awaited.

The bug causing the failure was the mock provider passed to the function
for testing not having the correct properties.