Workspace sync improvements
Makes workspace vault sync configurable,
so users can choose whether they want
workspace vaults to get synchronized,
skipped, or something inbetween.

Improves handling of non-repo vaults.
If some vaults are in repos but not others,
the non-repo vaults will simply be skipped.
Also fixed an issue where if a vault was not
in a repo, but the top-level Dendron
workspace was, the push command would
cause an exception.