fix: remove worddots ui flashing
This commit splits the `useMemo` for `groupedDailyNotes` and
`activeDate` into two separate calls. This allows these values to be
memoized independently and does not cross invalide eachother. The
visible effect in the UI was that the word dots flashed in/out after a
change inside a daily note. With this change the `groupedDailyNotes` value does not change when `activeDate` changes.