fix: clicking on a just deleted daily note
When the user deletes a daily note and clicks right after on the same
the an error was thrown in `CalendarView#handleReceiveMessage` with `Cannot read property 'fname' of undefined`.
This commit checks for the actual note and creates one if it does not exist.
The type return from `getEngine().notes[msg.data.id]` is therefore
missleading because it can be that the value returned is `undefined`.