feat(preview): update markdown on file save
I ad-hoc added a `notesRenderedTimestamp` redux state field for tracking the state the note was rendered and use it in the react view to rerender the note to markdown when the `updated` timestamp changed.
This can/should be probably done better.

Currently it has not the same responsiveness as with the "markdown preview enhanced" because when saving, first the engine needs to get updated and the `sync` call will therefore not catch that change.
The change will only be visible after the engine has been updated. For this, it looks to me, more structural changes first need to happen.

This commit does only make changes visible after save and **not** on text changes.