enhance(preview): add handlers to `fileWatcher`
This change was suggested by @kevinslin. I am not sure this is necessary because it seems this is already handled in `WorkspaceWatcher#onWillSaveTextDocument` and I am might miss something.

I was also considering moving `static refresh` to `workspace.ts` or some place higher since it looks like this refreshing note with messaging `onDidChangeActiveTextEditor` can be more context independent.