spike: nextjs based note preview (#843)
* spike: add initial note view

* spike: add clickhandler for anchore tags

* feat: add " Show Preview V2" command

* chore: use a more readable version naming event handlers

I have seen @kevinslin uses this style so I changed it.

* chore: remove trailing whitespace

* enhance: reveal existing panel

* feat(preview): open web links when clicking on them

* feat(preview): open local files when click on them

* spike(preview): add clarifying comments

* spike(preview): add todo comment

* fix(preview): open preview with note from last focused editor

* feat(preview): update markdown on file save

I ad-hoc added a `notesRenderedTimestamp` redux state field for tracking the state the note was rendered and use it in the react view to rerender the note to markdown when the `updated` timestamp changed.
This can/should be probably done better.

Currently it has not the same responsiveness as with the "markdown preview enhanced" because when saving, first the engine needs to get updated and the `sync` call will therefore not catch that change.
The change will only be visible after the engine has been updated. For this, it looks to me, more structural changes first need to happen.

This commit does only make changes visible after save and **not** on text changes.

* chore: simplify code by not inducing changes inside the redux store

It was unnecessary to change the structure of the redux store. Doing
this locally makes currently much more sense.

* fix(preview): use `contentHash` for checking note changes

`contentHash` seems to be the appropiate way to differ changes.

* chore(preview): use exiting handlers in `WorkspaceWatcher`

The existing `onWillSaveTextDocument` in `WorkspaceWatcher` has all necessary checks for actual changes.

* chore(preview): reuse existing functionality/pattern

* spike: add TODO comment

* spike: remove `console.log`

* enhance: rename handler to match semantically

* enhance(preview): add handlers to `fileWatcher`

This change was suggested by @kevinslin. I am not sure this is necessary because it seems this is already handled in `WorkspaceWatcher#onWillSaveTextDocument` and I am might miss something.

I was also considering moving `static refresh` to `workspace.ts` or some place higher since it looks like this refreshing note with messaging `onDidChangeActiveTextEditor` can be more context independent.

* feat(preview): update title when refreshing preview

* enhance(preview): preserve focus of current not when opening preview

Not sure if this is the behavior we want but that would be my
preference. I would want to keep writing of possible. To interact with
the preview you would anyway focus the window if e.g. you click on a
link.

* chore(preview): reuse exsiting functionality

* fix(preview): handle disposed preview

* chore(preview): remove unnecessary call

It's sufficient to call `ShowPreviewV2Command.onDidChangeHandler` in
`fileWatcher#onDidChange` only.

* fix(preview): error when opening preview without active editor

Opening the preview without having an open note or active editor made
the command crash. This allows the active editor to be `undefined`
because it can.