fix(calendar-note): correct note name from "Today" btn
The issue was that daily journal entries created with the "Today" button
had the format of `daily.journal.2021-07-02T12:18:57+02:00.md` (only in
month mode) which came to be because `onClickToday` had the oldest/first
reference of `onSelect` which renderd `defaultJournalDateFormat` `undefined`.

This adds the `onSelect` as a hook dependency. We probably should add https://www.npmjs.com/package/eslint-plugin-react-hooks to reduce risk of missing out on these.
This also adds a fallback if `defaultJournalDateFormat` is `undefined`.