enhance(calendar-view): not show a selected value in calendar when no active daily note exists
Before when opening the calendar without a active daily note opened
today's date would be selected. But since no active daily note is
present it should just not selected any date. Providing `null` prevents
that from happening.

Apparently this is not part of the typescript types but seems to work
internally nonetheless.