fix(calendar-note): correct note name from "Today" btn (#913)
* fix(calendar-note): correct note name from "Today" btn

The issue was that daily journal entries created with the "Today" button
had the format of `daily.journal.2021-07-02T12:18:57+02:00.md` (only in
month mode) which came to be because `onClickToday` had the oldest/first
reference of `onSelect` which renderd `defaultJournalDateFormat` `undefined`.

This adds the `onSelect` as a hook dependency. We probably should add https://www.npmjs.com/package/eslint-plugin-react-hooks to reduce risk of missing out on these.
This also adds a fallback if `defaultJournalDateFormat` is `undefined`.

* enhance(calendar-view): switch to `month` mode when clicking on "Today" btn

Clicking on the "Today" button shoud display the calendar in `month`
mode because that is what it implies.

* enhance(calendar-view): not show a selected value in calendar when no active daily note exists

Before when opening the calendar without a active daily note opened
today's date would be selected. But since no active daily note is
present it should just not selected any date. Providing `null` prevents
that from happening.

Apparently this is not part of the typescript types but seems to work
internally nonetheless.