fix(preview): show current active note when exec ShowPreview command
The issue has been reported this way:
- had a proj note open where I first tried out the preview V2, and it did switch back and forth between my journal and the proj note when I switched focus. I closed the preview and opened it back again on a different note but it will open up a preview of the proj note no matter what my current active window is.
   * Steps to reproduce:
      + Open preview on a note.
      + Close preview.
      + Switch to a different note.
      + Open preview again.

What was happening is that the first time the command was executed a
reference to the current active note was made and been reused everytime
the the preview has been reopened. The issue this implementation was
targeting seems to not be a problem anymore so I removed it.