feat: de-emphasize block anchors in the editor (#937)
* feat: de-emphasize block anchors in the editor

Parses the document to decorate block anchors. By parsing the note to provide these decorations, we ensure that only well-formed anchors will be highlighted, and the decorations will update automatically if we change anything about block anchors in parsing.

This also sets up an opportunity to decorate other things in the editor: for example we can highlight unlinked wikilinks, or color the alias of links differently etc. These highlights give the user immediate feedback, so I think it would be very useful.

* chore: test window decorations

* spike: use `expect`...`to` syntax for window decoration tests

* chore: fix merge conflict

Co-authored-by: Kevin Lin <kevinslin8@gmail.com>
Co-authored-by: Kevin <kevin@dendron.so>