fix: proc V4 misses options when converted from V5
The issue was that the proc in noteRef would create a new proc using V4, but it would get the `opts` from V5 proc. Since the opts of V4 and V5 procs don't match, this would break katex and mermaid in referenced notes, because without the correct options passed in these features are not active.

The only place where this happened, I guess, was the autocomplete & hover preview. Both use a V5 proc and use noteRef functionality to render the preview, which triggered the bad behavior.