feat: hashtag & reference autocomplete, and multi-vault autocomplete improvements (#991)
* chore: remove tag snippet

Now that we support hashtags, the tag snippet is redundant.

* feat: reference & hashtag autocomplete + xvault improvements

Provides autocomple for references `![[note]]` and hashtags `#my-tag`.

Displays the vault name next to the completion items.

When generating links, if there are notes with the same name in multiple vaults, then generates a cross-vault link `[[dendron-site://vault/note]]`.

 Also tried to display a preview of the suggested items. My implementation is good as far as I can figure out, but VSCode ignores the preview. Left this in for now.

 This includes all the improvements I had in mind for #887.

* chore: test improved autocomplete features

* spike: fix type errors

* fix: incompatible proc when rendering reference

* fix: proc V4 misses options when converted from V5

The issue was that the proc in noteRef would create a new proc using V4, but it would get the `opts` from V5 proc. Since the opts of V4 and V5 procs don't match, this would break katex and mermaid in referenced notes, because without the correct options passed in these features are not active.

The only place where this happened, I guess, was the autocomplete & hover preview. Both use a V5 proc and use noteRef functionality to render the preview, which triggered the bad behavior.

* chore: improve noteRef failure error message

* spike: disable note preview during autocomplete