feat(calendar-view): allow journal settings deviating from defaults (#1088)
* spike: replace hardcoded values

* chore: rename for better comprehension

To prefix terms with `default` implies there is an actual value apart from the default.
This is generally true but not from the calendar's perspective.

* feat(calendar-view): add support for custom `journal.dailyDomain` values

* feat(calendar-view): add support for custom `journal.journalName` values

* chore: remove unnecessary default value

* feat(calendar-view): add support for custom `journal.dateFormat` value

* chore: add explication comment

* chore(calendar-view): remove unnecessary checks

calendar-view currently does not take into account `journal.addBehavior`.

* chore: add `redux-logger` to `@dendronhq/common-frontend`

Since redux devtools are not avaiable inside vscode we can use a logger
to inspect redux actions order and data.

Eg: An error happening inside a redux thunk only revealed itself after
having access to the actual action object.

* fix: `syncChangedNote` call without a `note`

This can happen when no active editor is open and `ON_DID_CHANGE_ACTIVE_TEXT_EDITOR` is called with `note` being `undefined`.
By this `noteActive` (redux store) wil get set to `undefined` but prevents calling `syncNote`.

* fix: change type to allow to unset `noteActive`

When all active editors become close we should be able to set `noteActive` to `undefined`.

* chore: use default values from source of truth

* fix: revert import from `@dendronhq/engine-server`

This caused an error implying that this package cannot be used on the
frontend side. To use these default values we should propably move them
to `@dendronhq/common-all`.

* chore: add explication comment

* chore: make calendar-view run without `config`

This makes the calendar-view more resilient since it still runs without
available `config`. Meaning it will show a calendar that is empty and
also won't create entries when clicking on a day.

* fix: consider notes length in calendar-view's memo `areEqual` function

The reason we need to consider length rather only existence is because I
observed that the order of redux action can land into a state where
`engine.notes` can contain first only the current active editor note,
being triggered by `syncNote`.
At a later timestamp the `initNotes` fulfilles and adds the rest of the notes.

* fix: remove wrongly set condition