fix: prevent throwing error
Patching only the relevant functions to not trow if `date` is not defined.

These two functions are called from the `Header` component of `antd`'s `Calendar` component.
If no active editor is open then `value` is not defined. In this state
selected a month/year in the header threw an error.

This is a quickfix solution.