spike: calc tree data immediately
`MenuView` needs `expandKeys` to have the right values on mount.
Because it won't expand keys on a second render when they change.

This commit does this by moving the calculation out of the `useEffect`
which is at this point in time not necessary. If the calculations are
expensive `useMemo` is a better fit.