Revert "spike: make `Sider` full the whole document height"
This reverts commit cb7103f9fd70379570dbf71b2a35f5ea84b1f496.

Reason that the unwanted background color happened was not because if
sidebar not having the right height (which actually comes from use `vh`)
but from content overflowing the main area. A youtube video was it in
this case. So to solve this we need to have a max width to embeded media
content.