spike: add hover styles
The intend is to add further indicators for what the click might do.
The underline should create an expectation to the user that the page
will be opened. The hover background for the expand icon should indicate
that an interactive button is there, which ideally should indicate that
this would not open a new page but only expand the submenu.