fix: solve submenu popups flashs
There is a bug in ant's `Menu` where switch the prop `openKeys` to undefined when collapsing the sider and back where the submenu popups
are greated when they should not be. This creates a flashy appearance.
Using `key` prop to unmount/mount the component when switching collapsable state solves this because `Menu` is being rendered with a clear state and no "knowledge" of what previously happend. This might have some performance caused depending on how the internals for `Menu` work.