enhance: Nextjs template style improvements (#1193)
* fix: adjust sidebar height

This places the sidear in its correct location so that it is fully
visible. Before a part of the bottom was cut of because and therefore
hid the horizontal sidebar.

* spike: seperate layout components from content components

I think its more predictable to have layout and content components not
merged. In this case it means that I extracted the antd's `Footer` out
into `_app.tsx`.

* fix: correct LogoOrTitle clickable area

Before the area expanded to the right until the lookup bar.

* chore: remove unnecessary css `fixed` properties

The header is already fixed. Having the logo also fixed should not have
the behavior.

* spike: simplify title rendering

* spike: make titleOrLogo adapt to parent height

Less hard-coded and self-adjusting behaviors values are better.

* fix: make logo behave fluid

In small screensizes the logo overflows into the lookup bar.

* spike: cleanup style constants

* spike: experimantel V2 layout component

* spike: implement tree view with antd's `Menu` component

* spike: use `onClick` from `Menu`

* spike: calc tree data immediately

`MenuView` needs `expandKeys` to have the right values on mount.
Because it won't expand keys on a second render when they change.

This commit does this by moving the calculation out of the `useEffect`
which is at this point in time not necessary. If the calculations are
expensive `useMemo` is a better fit.

* spike: use `Menu`'s selection highlighting

* spike: set active note on menu expansion

* spike: reduce intent level

* spike: align `DendronTreeView` in sider

* spike: directly change activeNotes on page change

A router page change does not trigger a re-rerender.

* spike: comment out old `DendronLayout` component

* spike: set viewport metatag

* spike: differentiate between onExpand and onSelect on SubMenu items.

* spike: use `Menu` as an uncontrolled component

This allows to close the menu even then `expandedKeys` says otherwise.

* spike: use custom open/close icon

There is still an issue when using a custom expandIcon. When using the
Menu in a collapsed state some parents don't open on click.

* spike: start implement open styles for submenu items

* spike: show `Sider` trigger button

* spike: use `Menu` as a controlled component when not it mobile view

* spike: track submenus open/close state

* spike: add selected submenu styles

* spike: adjust expand submenu logic

This prevents unnecessary calls from `onOpenChange`.

* spike: cleanup

* spike: extract style constant

* spike: add icons

* spike: make menu popups work with touch gestures

The issue was that with touch gestures the submenus would not open.
This happened because on the icon component was unmounted/mounted and
probably lost thereof registered event handlers.

* spike: adjust footer padding

* spike: cleanup style constants

* spike: rename component

I think the new name communicates more the case that it is the **menu**
for the dendron site.

* spike: make `Sider` full the whole document height

Since the `Sider` has a slight darker background this would otherwise result in the area after the sider to be white.
We want the whole sidebar area to have the same color.

* spike: update the color of the menu expand button to dendron accent color

* spike: extend submenu open/close touch area

* Revert "spike: make `Sider` full the whole document height"

This reverts commit cb7103f9fd70379570dbf71b2a35f5ea84b1f496.

Reason that the unwanted background color happened was not because if
sidebar not having the right height (which actually comes from use `vh`)
but from content overflowing the main area. A youtube video was it in
this case. So to solve this we need to have a max width to embeded media
content.

* spike: move inline styles into scss

This solves an issue of styles being unwanted overritten.
Further this allows to put hover styles for the split button situation.
Though this might be not a solution we want https://www.nngroup.com/articles/split-buttons-navigation/

* spike: make open/close submenu icons position more aligned

By rendering two seperate buttons instead of using css rotation they
stay on-changing direction more or less on the same position.

Though this is not the best for render performance and be probably
should do this only within css.

* spike: add hover styles

The intend is to add further indicators for what the click might do.
The underline should create an expectation to the user that the page
will be opened. The hover background for the expand icon should indicate
that an interactive button is there, which ideally should indicate that
this would not open a new page but only expand the submenu.

* spike: adjust icon for collapsed menu

* spike: use next/link for navigating to homepage

* fix: solve submenu popups flashs

There is a bug in ant's `Menu` where switch the prop `openKeys` to undefined when collapsing the sider and back where the submenu popups
are greated when they should not be. This creates a flashy appearance.
Using `key` prop to unmount/mount the component when switching collapsable state solves this because `Menu` is being rendered with a clear state and no "knowledge" of what previously happend. This might have some performance caused depending on how the internals for `Menu` work.