spike: change layout technique for easier sidebar coloring
This commit builds the centered layout not by using the `margin: 0 auto`
approach but rather using flexbox. This way the we have two area
(sidebar, main) which has the advantage that they can have different properties (backgroundColor,
etc.) seperately. To create a centered appearance this sets a
`paddingLeft` value on the sidebar based on the half of the leftover
from the substraction of viewport width and the `lg` breakpoint. This
aligns the sidebar on the left where the centered content would start.
The main content area flows to the right and is caped by a max-width.