enhance: FM tags highlighting & missing tag notes highlighted as missing
Frontmatter tags are now highlighted in the same way hashtags are.
Hashtags and frontmatter tags with missing tag notes are now
highlighted in the same way wikilinks for missing notes are.
The idea is that since a lot of features (autocomplete, backlinks in publishing etc.)
depend on the tag notes actually existing, the highlighting would encourage users to create the tag notes.

![](https://i.imgur.com/LyAM9MR.png)

![](https://i.imgur.com/KirZPnL.png)

![](https://i.imgur.com/LJ4eguI.png)

![](https://i.imgur.com/C1Vyn1l.png)