fix: block anchors attached to code blocks in publishing (#1267)
* fix: block anchors attached to code blocks in publishing

* fix: wrap top-level anchor to avoid on hover of all content

Without wrapping, the top-level anchors we install for code blocks would always appear wherever the user was hovering over on the page. Now it will only appear when the user hovers over where the anchor should be. This makes them less discoverable on published pages, but I can't find a better solution.

* spike: fix bad imports