fix: disable certain decorations for long notes to avoid performance hit
The slowdown when editing files that have a lot of wikilinks or hashtags were caused by some expensive decorations which use `getNotesByFname` many times. `getNotesByFname` scans all notes linearly, and is expensive.

The fix here is a workaround that avoids the problematic decorations for long notes. The decorations disabled are the colors of hashtags (hashtag color squares show up empty), and the broken wikilink highlighting (all wikilinks show up as existing/valid).