fix: possible error if open note is changed quickly after edit
This issue hasn't been reported by any users thanks to the check in `WorkspaceWatcher`,
but it does mean that commands that edit the current note and then open a new note don't trigger `onDidChange`. We had to work around this in the selection to link lookup modifier, it's best to fix it for future proofing.