chore(build): upgrade to typescript 4.4.4 and add `dendron-plugin-views` package
This commit adds [[Dendron Plugin Views|dendron://dendron.dendron-site/pkg.dendron-plugin-views]] which is used to generate webviews for vscode
To deal with typescript issues that arose due to this package, we ended up also upgrading the version of typescript used in our `yarn.lock` file to the latest (4.4.4) which results in stricter type checking.

- type checking is more strict
- error type now has to be cast as `unknown` or `any`, see [useUnknownInCatchVariables](https://www.typescriptlang.org/tsconfig#useUnknownInCatchVariables)

Note that as part of this commit, we did not actually address the `4.4.4` error checks but instead cast `error: any` and `ts-ignore` to many of the issues.
This commit is to do the initial upgrade - fixing the existing type issues will be saved for a later commit