feat: Initialize Workspace command can create native workspaces
If the user currently has a non-Dendron workspace open, we now offer to initialize that as a Dendron native workspace.

Fixed a bug with handling existing roots where cancelling the prompt would create a workspace anyway, and also improved the user-friendliness of it.

Finally, also fixed a bug where non-Dendron workspaces are recognized as Native Dendron workspaces.

![](https://i.imgur.com/c5o04yl.png)
![](https://i.imgur.com/A2PkgVo.png)
![](https://i.imgur.com/PzUIT7M.png)