fix(publish): make mermaid work consistently on published sites
Mermaid diagrams would only render on first page load but not upon subsequent page navigation. This makes mermaid always render when enabled.

Note that this also changes the signature for mermaid rendering in `dendron-plugin-views` as we're pushing down the mermaid rendering into `DendronNote` which is shared by `nextjs-template` and `dendron-plugin-views`.

There's also an issue where `getProps(config, 'preview')` was returning `undefined` in the browser. It works fine locally (added a test for this as well).