feat: better note decorator bad performance handling
We now give up if calculating decorations takes too long instead of doing it based on the note length, which is a better experience when dealing with notes that are long but don't have a lot of links in them.

Also added a debounce to the decorator updates to avoid the editor getting bogged down with a lot of async operations that haven't yet completed, which then blocks other async functions like note lookups.