enhance(cli): cli uses separate port file (#1766)
* enhance: cli write to own path

* enhance(cli): do not overwrite ws dendron port when activating CLI

The Dendron CLI will create its own instance of an API server for most commands. This engine, when spawned, will write a port file at `.dendron.port` so that other processes can discover the port that the engine server is running on.
The issue is that a running workspace has its own port file which will be overridden. This will make commands like `dendron {cmd} --attach` fail as the port is no longer accurate.

This change causes the dendron-cli to write the engine port at `dendron.port.cli` instead.