enhance(publish): attempt to update template in-place
Verify that the repository is using the expected remotes, the correct
branch is checked out, the branch tracks the correct remote branch and
that the working tree is clean.

If it is, issue a git fetch followed by git reset.

If any of the checks fail, we fall back to removing the .next directory
and trying again.