use nonegate to normalize how patterns work
The way we build patterns, a pattern that starts with `!(thing)` means "anything that's not `thing`".
However, if the same pattern is used for the top level (schema part with `root` as the parent),
then the final pattern becomes
`!(thing)/...`. Because the pattern starts with `!`, this means the entire pattern is negated.

By adding the option `nonegate` we disable this behavior, so `!(thing)` works the same way no matter which part of the schema it is used in.