chore(refactor): Migrate backlinkspanel to use EngineEventsEmitter
Instead of refreshing the backlinkspanel manually from different sources, we want the backlinks panel to subscribe to an event and refresh upon event fire.
For this panel, it will refresh upon a note change from the engine.

This change will help address cyclic dependencies as well as misc tech debt