fix(publish): renders consitent layout on mobile and non-mobile (#3272)
* fix(publish): render DendronLayout by default (mobile-first) vertically

On mobile the layout's children should be rendered vertically (meaning:
`flex-direction: column` which is the default). This will solve related issue #2175 but will introduce a new problem.
In my view this new problem is the actual issue we are facing. See next
commit.

* fix(publish): make sidebar and content render horizontal on non-mobile sceens

This commit follows the previous commit and makes use of antd [sidebar feature](https://ant.design/components/layout/#components-layout-demo-side).
It works the following way:

- A `Layout` component gets the css class `.ant-layout-has-sider` applied if it has a `Slider` somewhere nested as a child and is not collapsed.
- `.ant-layout-has-sider` applies the css property `flow-direction: row`
  which results in the sidebar and content to be rendered horizontally.

Lets look at this with an example.

Here the `Layout#hier` component gets the css class `.ant-layout-has-sider` applied if the `Slider` is not collapsed.

```ts
<Layout>
  <Header>Header</Header>
  <Layout id="hier">
    <Sider>Sider</Sider>
    <Content>Content</Content>
  </Layout>
  <Footer>Footer</Footer>
</Layout>
```

In this example only `Layout#hier-2` recieves this css class.
This would result in `Layout#hier` and `Content` to still stay
vertically oriented.

```ts
<Layout>
  <Header>Header</Header>
  <Layout id="hier">
    <Layout id="hier-2">
      <Sider>Sider</Sider>
    </Layout>
    <Content>Content</Content>
  </Layout>
  <Footer>Footer</Footer>
</Layout>
```

This commit basically removed the `Layout#hier-2` by replacing it with a `div` which makes the
`.ant-layout-has-slider` css class be applied at the right level.

This is related to #2175 and only established the groundwork for the
actual fix. See next commit.

* chore(publish): remove unused `BasicLayout` component