chore(views): remove premature communication channel
Currently preview lock implements the behavior in which the lock state
gets reseted to be unlocked when disposing the PreviewPanel. Therefore
opening the PreviewPanel will also start with unlocked. If the
state-lock outlives that PreviewPanel's webview panel we would need to
sync it with the again visible webview panel.
This could be done in `packages/plugin-core/src/components/views/PreviewPanel.ts 148`