fix(views): make lock from preview work
This only points to the pattern that fixes the lock from preview because
the note reference that is being added to the lock message type only
contains the filename and not the filepath. Inside PreviewPanel we make
the lock state using the actual note fileName as the whole path. Next
commit will solve this using note id.