enhance(views): toggle preview lock from inside preview (#3327)
* enhance(views): allow toggle previewlock from inside preview

* fix(views): make lock from preview work

This only points to the pattern that fixes the lock from preview because
the note reference that is being added to the lock message type only
contains the filename and not the filepath. Inside PreviewPanel we make
the lock state using the actual note fileName as the whole path. Next
commit will solve this using note id.

* chore(views): use noteId for preview lock

The client provides `props.ide.noteActive?.fname` which does not contain
the whole filepath rather only the filename. We need a gui and noteId
provides that also.

* enhance(views): makes preview's unlocked icon less salient