fix(views): show whitespace for links in headers
For the curious here an explanation to the issue:

The 'white-space' processing model rules get applied to lines. A line is a set of inline elements.

One of the rules is:

> As each line is laid out, […] If a space (U+0020) at the end of a line has ‘white-space’ set to ‘normal’, ‘nowrap’, or ‘pre-line’, it is also removed.

A line of inline elements might get its whitespace at the end removed (NOTE: The line and not each inline element).

If we give a container `flex:display` the children will become block elements. This means that each child counts as its own line from the POV of the 'white-space' processing model.
Each flex-child will get its whitespace at the end removed.
The prevent that from happening we can apply the css property: `white-space: pre-wrap;`. `pre` means sequences of white space are preserved and tells the 'white-space' processing model to not remove that last whitespace.

- https://www.mattstobbs.com/flexbox-removing-trailing-whitespace/
- [white space processing model](https://www.w3.org/TR/CSS21/text.html#white-space-model)
- https://developer.mozilla.org/en-US/docs/Web/CSS/white-space