wip(publish, sidebar): add base implementation for sidebarconfig parsing
This commit adds two libraries for doing the work. `@badrap/valita` for
typesafe validation and parsing and `neverthrow` to allow working with
[Result type](https://en.wikipedia.org/wiki/Result_type).

The `parse` file provides a general way to parse valita schemas and get
a `Result` type returned.

NOTE: first I used [zod](https://github.com/colinhacks/zod) but with
valita we get a smaller and faster solution and the issue I stumbled on
with zod, which is combining recursive types with discoint union types,
works in valita.