fix(publish, sidebars): use `getAllParents` based on `tree.json`
There where two `getAllParents` functions with the difference (apart
from one including the referencing note) that one
operates on the whole notes lists and the other only notes existing in
tree data (`tree.json`). This comment changes the uses to the one using
tree data and deletes the other since it is not used anymore.