feat(publish): adds configuration for sidebar (#3448)
* chore: remove unused import

* wip(publish, sidebar): load sidebar config

* wip(publish, sidebar): add base implementation for sidebarconfig parsing

This commit adds two libraries for doing the work. `@badrap/valita` for
typesafe validation and parsing and `neverthrow` to allow working with
[Result type](https://en.wikipedia.org/wiki/Result_type).

The `parse` file provides a general way to parse valita schemas and get
a `Result` type returned.

NOTE: first I used [zod](https://github.com/colinhacks/zod) but with
valita we get a smaller and faster solution and the issue I stumbled on
with zod, which is combining recursive types with discoint union types,
works in valita.

* fix(publish, sidebar): add missing sidebarPath config

* fix: remove accidentally added merge conflict markers

* chore: update typescript to 4.7.x

This allows to have acces to [`at()`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/at)

* chore(common): enable `noUncheckedIndexedAccess`

This is a good check to have.

* chore: cleanup

* wip(publish, sidebar): lays out boilerplace for sidebars processing

Contains the beginnings of unwrapping "autogenerated" items.

* fix(publish, sidebar): ts error my removing `sidebarItemsGenerator`

The error raised is TS4023.

* wip(publish, sidebar): use `root` keyword for toplevel hierarchySource

* fix(publish, sidebar): move requiring of sidebar config file into server runtime

`common-all` code gets imported in client side code and there we don't
have access to the filesystem. this commit extracts that part out.

* wip(publish, sidebar): use new sidebarConfig to generate a `TreeMenu`

* feat(publish, sidebar): convert autogenerated sidebarItem to sidebar slice

* refactor(publish, sidebar): extract idSchema

* feat(publish, sidebar): make autogenerated work from root

* fix(publish, sidebars): use `getAllParents` based on `tree.json`

There where two `getAllParents` functions with the difference (apart
from one including the referencing note) that one
operates on the whole notes lists and the other only notes existing in
tree data (`tree.json`). This comment changes the uses to the one using
tree data and deletes the other since it is not used anymore.

* refactor(publish): make `getAllParents` only return parents

* refactor: remove unused `note2TreeDatanote` function

* chore(publish, sidebar): and skeleton for possible unit-tests

* refactor: move function closer to its actual bounded context.

- [Bounded context](http://ddd.fed.wiki.org/view/bounded-context)

* feat(publish, sidebars): add sorting

This sorts be default alphabetically based on the filename and
additionally by `nav_order` frontmatter field if set.

* feat(publish,sidebars): use `sort_order` config field

* refactor(sidebars): rename and use pipe pattern

* feat(publish, sidebar): compute breadcrumps from `tree.json`

`tree.json` is the output that has embeded the sidebar config so it
should the the source for building the breadcrumb.

* fix(publish, sidebar): ignore `root` note when computing top hierachies

* feat(publish, sidebar): error when sidebarItem's associated note cannot be found

* chore(publish, sidebar): resolve sidebarItem identifier

The sidebarItem identifier is captured by the `id` field.
It can point either to a note `id` or a note `fname`.
Duplicate `fname` entries are resolved by `duplicateNoteBehavior` config option.

* chore(publish, sidebar): cleanup types

* fix(publish, sidebars): ignore root keyword when resolving item id

* enhance(publish, sidebar): add an error for invalid sidebarItem

* fix(publish, sidebars): remove accidentally added code

* refactor(publish, sidebar): encapsulate general result type in a descriptive dendron type

See https://github.com/supermacro/neverthrow/wiki/Introduction:-Type-Safe-Errors-in-JS-&-TypeScript-(10-minute-read)#making-your-types-more-intuitive

* chore: mark test as todo

* chore: remove accidentally added code

* refactor(publish, sidebars): cleanup lefovers

* chore: describe result type

* fix(publish, sidebar): treat `null` value for `sidebarFilePath`

* chore(typescript): make "es2022.array" features usable

* chore: remove nonexisting file import

* chore: remove unused import

* refactor(publish, sidebars): simplify

* chore: temporarily disable tests

* feat(publish, sidebar): change root keyword to `root`

* chore: add test-workspace export helper to Makefile

* feat(publish, sidebar): remove support for multiple sidebars

* chore: remove es2022 `.at` array access usages

For now rely on older methodes.

* refactor: replace `Result` type with `RespV3`

* chore: rename sidebars to sidebar

I missed these renamings in the previous commit when removing multiple
sidebar logic.

* chore(publish, sidebar): add description

* chore: rename make target to be more descriptive

* chore(publishing): finish setting up new `sidebarPath` config option

* chore: remove hardcoded error message

* refactor: remove experimental commmon util function

* chore: fix import path

* refactor(publish, sidebar): switch parsin library back to `zod`

* refactor: follow tests placement convention

* feat(publish, sidebar): use `nav_exclude_children` and `fm.has_collection` in `generateTreeData`

* chore: convert treeUtil tests to new implementation

* chore(publish, sidebar): add tests

* chore(test, common-all): refactor test to be closer to original state

Original meaning how it was before the new sidebar.

* chore(publish, sidebar): reimplement missed custom validation logic in `zod`

* chore(test): fix tests

* test(sidebar): remove unnecessary snapshot test

* test(publish, sidebar): remove redundant snapshot test

A sibling snapshot test basically tests already the same thing.

* test(publish, sidebar): add behavior assertions

* feat(publish, sidebar): throw when provided sidebar file doe snot exist

* enhance(publish, sidebar): fail fast for `sidebarPath` check

`SiteUtils.filterByConfig` takes time and letting a throw when a provided `sidebarPath` could not be found happen before that saves the user some time.

* test(publish, sidebar): add test for custom sidebar

* test(publish, sidebar): add test for autogenerated sidebar slice

* refactor(publish, sidebar): move logic into `sidebar.ts`

Logic that dictates would should be displayed in the sidebar should reside inside sidebar.ts

* refactor(publish, sidebar): move `nav_exclude` logic into `sidebar.ts`

Logic that dictates would should be displayed in the sidebar is best placed inside sidebar.ts.

* chore(publish, sidebar): add reference to docs

* test(publish, NextjsExportPod): convert to given-when-then style

* test(publish, NextjsExportPod, sidebarPath): add unit tests

* feat(publish, sidebar): improve error messages

* fix(publish, sidebar): make function return RespV3 instead of throwing

* enhance(publish, sidebar): fail earlier when sidebarConfig parsing failed

This commit puts sidebarConfig parsing before calculation notes
dictionary which is expensive and takes time. This way resources are
saved and the user gets a faster response.