fix(publish): align heading anchors
This changes the position of the heading anchors from the left to the
right. This makes the alignment code much simpler and less fragil. From
the users POV having the anchor in the right should not effect UX. The
same approach is taking by others: https://nextra.vercel.app/features/mdx#blockquote.